renamed from camelCase to dashes
