+++ packages/main/src/api/controllers/file-system-controller.ts
@@ -1,6 +1,6 @@
 import fs from "fs";
-import formatDate from "../helpers/formatDate";
-import type htmlData from "../types/htmlData";
+import formatDate from "../helpers/format-date";
+import type htmlData from "../types/html-data";
 
 // TODO:
 // Return array of all files in a directory
