+++ packages/main/src/api/helpers/format-date.ts
